(PC-28702) refactor(Accordion): Remove switch prop and use leftComponent instead
